requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.11.0